\name{QCqPCR-package}
\alias{QCqPCR-package}
\alias{QCqPCR}
\docType{package}
\title{Functions for quality control of qPCR data}
\description{
The package provides functions for quality control of RT-qPCR data.
}
\details{
\tabular{ll}{
Package: \tab QCqPCR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-10-20\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
library(QCqPCR)
}
\author{
James Perkins <jperkins@biochem.ucl.ac.uk>\cr
Matthias Kohl <Matthias.Kohl@stamats.de>

Maintainer: James Perkins <jperkins@biochem.ucl.ac.uk>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{ package }
