\name{BoxWhisker}
\alias{BoxWhisker}
\alias{qPCRBoxWhisker}
\alias{qPCRBoxWhisker,qPCRBatch-method}
\title{ Box and whisker plots }
\description{ 
  yet to fill 
}
\usage{
qPCRBoxWhisker(qPCRBatch, \dots)

\S4method{qPCRBoxWhisker}{qPCRBatch}(qPCRBatch, cutOff, statType, plotType) 
}
\arguments{
  \item{qPCRBatch}{ data set containing data on different Ct values normed to the different HKGs }
  \item{\dots}{ Extra arguments, detailed below }
  \item{cutOff}{ yet to fill }
  \item{statType}{ yet to fill }
  \item{plotType}{ yet to fill }
}
\details{ aaa
}
\section{Methods}{
  \describe{
    \item{qPCRBoxWhisker}{(qPCRBatch): method for object of class \code{"qPCRBatch"} }
  }
}

\value{  yet to fill  }
%\references{ ~put references to the literature/web site here ~ }
\author{ James Perkins \email{jperkins@biochem.ucl.ac.uk} }
%\examples{
%}
\keyword{classes}
