\name{BoxWhisker}
\alias{BoxWhisker}
\alias{qPCRBoxWhisker}
\alias{qPCRBoxWhisker,qPCRBatch-method}
\title{ Box and whisker plots }
\description{ aaa
}
\usage{
qPCRBoxWhisker(qPCRBatch, cutOff, statType, plotType)
}
\arguments{
    \item{qPCRBatch}{ data set containing data on different Ct values normed to the different HKGs }
    \item{cutOff}{}
    \item{statType}{}
    \item{plotType}{}
}
\details{ aaa
}
\section{Methods}{
  \describe{
    \item{qPCRBoxWhisker}{(qPCRBatch): ... }
  }
}

\value{ aaa }
%\references{ ~put references to the literature/web site here ~ }
\author{ James Perkins \email{jperkins@biochem.ucl.ac.uk} }
%\examples{
%}
\keyword{classes}
